Add static keyword to RotationBase::exponentialMap(vector)
The method does not modify the object in any way, but returns the exponential map for the given vector.
So it can be static.

Without this patch at least clang 3.8 emits a warnings due to use of uninitialized variables when used like:

{
  kindr::RotationQuaternionPD q = q.exponentialMap(theta);
  // ==> error: variable 'q' is uninitialized when used within its own initialization [-Werror,-Wuninitialized]
}

which seems reasonable. Ideally those lines would be replaced by something like

{
  auto q = kindr::RotationQuaternionPD::exponentialMap(theta);
  // ...
}